Pull out ChoiceNameFunctionHandler
